Moves accordion-item-title to wrapper file structure
