+++ src/AccordionItem/accordion-item.spec.js
@@ -4,7 +4,7 @@ import React from 'react';
 import { mount } from 'enzyme';
 import renderer from 'react-test-renderer';
 import { Provider } from 'unstated';
-import AccordionItemTitle from '../AccordionItemTitle/accordion-item-title';
+import AccordionItemTitle from '../AccordionItemTitle/accordion-item-title-wrapper';
 import AccordionItemBody from '../AccordionItemBody/accordion-item-body-wrapper';
 import AccordionItem from './accordion-item-wrapper';
 import { resetNextUuid } from '../ItemContainer/ItemContainer';
