+++ src/index.js
@@ -3,7 +3,7 @@
 // eslint-disable-next-line
 import Accordion from './Accordion/accordion-wrapper';
 import AccordionItem from './AccordionItem/accordion-item-wrapper';
-import AccordionItemTitle from './AccordionItemTitle/accordion-item-title';
+import AccordionItemTitle from './AccordionItemTitle/accordion-item-title-wrapper';
 import AccordionItemBody from './AccordionItemBody/accordion-item-body-wrapper';
 import { resetNextUuid } from './ItemContainer/ItemContainer';
